+++ src/routes/(app)/administration/send-logs/javascript/+page.server.ts
@@ -5,11 +5,11 @@ import type { NodeLogFlavor } from '$lib/types';
 import type { PageServerLoad } from './$types';
 
 const OTEL_INSTALL =
-	'npm install @opentelemetry/api-logs @opentelemetry/sdk-logs @opentelemetry/exporter-logs-otlp-http';
+	'npm install @opentelemetry/api-logs @opentelemetry/sdk-logs @opentelemetry/exporter-logs-otlp-proto';
 
 const OTEL_EXAMPLE = `import { logs } from '@opentelemetry/api-logs';
 import { LoggerProvider, BatchLogRecordProcessor } from '@opentelemetry/sdk-logs';
-import { OTLPLogExporter } from '@opentelemetry/exporter-logs-otlp-http';
+import { OTLPLogExporter } from '@opentelemetry/exporter-logs-otlp-proto';
 
 const provider = new LoggerProvider({
     processors: [new BatchLogRecordProcessor(new OTLPLogExporter())]
@@ -31,11 +31,11 @@ const log = pino(transport);
 log.info('Hello from Pino to Logwiz');`;
 
 const WINSTON_INSTALL =
-	'npm install winston @opentelemetry/api-logs @opentelemetry/sdk-logs @opentelemetry/exporter-logs-otlp-http @opentelemetry/winston-transport';
+	'npm install winston @opentelemetry/api-logs @opentelemetry/sdk-logs @opentelemetry/exporter-logs-otlp-proto @opentelemetry/winston-transport';
 
 const WINSTON_EXAMPLE = `import { logs } from '@opentelemetry/api-logs';
 import { LoggerProvider, BatchLogRecordProcessor } from '@opentelemetry/sdk-logs';
-import { OTLPLogExporter } from '@opentelemetry/exporter-logs-otlp-http';
+import { OTLPLogExporter } from '@opentelemetry/exporter-logs-otlp-proto';
 import { OpenTelemetryTransportV3 } from '@opentelemetry/winston-transport';
 import winston from 'winston';